requirements.txt
aiohttp==3.9.3
aiohttp_apispec==3.0.0b2
aiohttp_cors==0.7.0
aiohttp_session==2.12.0
alembic==1.13.1
asyncpg==0.29.0
cryptography==42.0.5
greenlet==3.0.3
marshmallow==3.21.0
pytest==8.0.2
pytest-aiohttp==1.0.5
pytest-asyncio==0.23.5
PyYAML==6.0.1
ruff==0.4.2
SQLAlchemy==2.0.27
python-dotenv==1.0.1

aiosignal==1.3.1
apispec==6.6.0
attrs==23.2.0
cffi==1.16.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pluggy==1.4.0
pycparser==2.21
typing_extensions==4.10.0
webargs==8.4.0
yarl==1.9.4
